+++ demos/employees/src/main/java/is/codion/demos/employees/ui/EmployeeEditPanel.java
@@ -73,7 +73,7 @@ public class EmployeeEditPanel extends EntityEditPanel {
 		EntityComboBox departmentBox = createComboBox(Employee.DEPARTMENT_FK).build();
 		NumberField<Integer> departmentNumberField = departmentBox.integerSelectorField(Department.DEPARTMENT_NO)
 						.transferFocusOnEnter(true)
-						.onBuild(field -> addValidator(Employee.DEPARTMENT, field))
+						.validIndicator(editModel().editor().value(Employee.DEPARTMENT_FK).valid())
 						.build();
 
 		component(Employee.DEPARTMENT_FK).set(departmentNumberField);