+++ demos/petclinic/src/main/java/is/codion/demos/petclinic/ui/PetclinicAppPanel.java
@@ -31,6 +31,7 @@ import is.codion.demos.petclinic.model.PetclinicAppModel;
 import is.codion.demos.petclinic.model.VetSpecialtyEditModel;
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.plugin.flatlaf.intellij.themes.arc.Arc;
+import is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityApplicationPanel;
 import is.codion.swing.framework.ui.EntityPanel;
@@ -120,6 +121,7 @@ public final class PetclinicAppPanel extends EntityApplicationPanel<PetclinicApp
 		Locale.setDefault(new Locale("en", "EN"));
 		ReferentialIntegrityErrorHandling.REFERENTIAL_INTEGRITY_ERROR_HANDLING
 						.set(ReferentialIntegrityErrorHandling.DISPLAY_DEPENDENCIES);
+		ValidIndicatorFactory.FACTORY_CLASS.set("is.codion.plugin.flatlaf.indicator.FlatLafValidIndicatorFactory");
 		EntityApplicationPanel.builder(PetclinicAppModel.class, PetclinicAppPanel.class)
 						.applicationName("Petclinic")
 						.domainType(Petclinic.DOMAIN)