+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComboBoxPanel.java
@@ -18,10 +18,12 @@
  */
 package is.codion.swing.framework.ui.component;
 
+import is.codion.common.state.ObservableState;
 import is.codion.common.value.Value;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.swing.common.ui.component.builder.AbstractComponentBuilder;
 import is.codion.swing.common.ui.component.builder.ComponentBuilder;
+import is.codion.swing.common.ui.component.indicator.ValidIndicatorFactory;
 import is.codion.swing.common.ui.component.value.AbstractComponentValue;
 import is.codion.swing.common.ui.component.value.ComponentValue;
 import is.codion.swing.common.ui.key.TransferFocusOnEnter;
@@ -234,6 +236,12 @@ public final class EntityComboBoxPanel extends JPanel {
 		protected void enableTransferFocusOnEnter(EntityComboBoxPanel component) {
 			TransferFocusOnEnter.enable(component.comboBox());
 			component.buttons.forEach(TransferFocusOnEnter::enable);
+		}
+
+		@Override
+		protected void enableValidIndicator(ValidIndicatorFactory validIndicatorFactory,
+																				EntityComboBoxPanel component, ObservableState valid) {
+			validIndicatorFactory.enable(component.comboBox, valid);
 		}
 
 		private EntityComboBox createComboBox() {