+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -26,6 +26,7 @@
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#USER_PREFERENCES_ENABLED}
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#RESTORE_DEFAULT_PREFERENCES}
  * <li>{@link is.codion.swing.framework.ui.EntityApplicationPanel#CALL_SYSTEM_EXIT}
+ * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#VALID_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#MODIFIED_INDICATOR}
  * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#MODIFIED_INDICATOR_UNDERLINE_STYLE}
  * <li>{@link is.codion.swing.framework.ui.EntityEditComponentPanel#DEFAULT_TEXT_FIELD_COLUMNS}